+++ src/Backend/Component/Core/ConfigurationStep/General.php
@@ -225,8 +225,6 @@ class General extends ConfigurationStep
         ->setSgOwnerDomain(\Contao\Environment::get('base'))
         ->setSgOwnerHost(CoreConfig::DEFAULT_OWNER_HOST)
         ->setRejectLargeUploads(true)
-        ->setFileusageSkipReplaceInsertTags(true)
-        ->setFileusageSkipDatabase(true)
         ;
 
         $this->localConfigManager->save($config);